+++ package.json
@@ -5,14 +5,16 @@
   "repository": "git@github.com:bedstuystrong/functions-2.git",
   "author": "bedstuystrong@bedstuystrong.com",
   "license": "MIT",
+  "type": "module",
   "dependencies": {
-    "@netlify/functions": "^1.6.0",
+    "@netlify/functions": "^2.3.0",
     "@netlify/integrations": "^0.6.0",
     "@netlify/sentry": "^0.0.8",
     "@sendgrid/mail": "^7.7.0",
     "@sentry/node": "^7.75.1",
     "airtable": "^0.12.1",
     "busboy": "^1.6.0",
+    "eta": "^3.1.1",
     "lodash": "^4.17.21",
     "mailparser": "^3.6.5",
     "tiny-invariant": "^1.3.1"
